pyproject.toml
[tool.poetry]
name = "SUP"
version = "0.1.0"
description = ""
authors = ["Kazmiryk Ruslan <callogan217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
argon2-cffi = "^23.1.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
json-log-formatter = "^0.5.2"
python-json-logger = "^2.0.7"
aiosmtplib = "^3.0.1"
pre-commit = "^3.7.0"
typer = "^0.12.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
ruff = "^0.1.14"
mypy = "^1.8.0"

[tool.ruff]
line-length = 120
indent-width = 4
exclude = ["./migrations", "./tests", "__init__.py"]

[tool.ruff.lint.pycodestyle]
max-line-length = 120

[tool.ruff.lint]
extend-ignore = ["F821"]
select = ["F", "E", "I"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
